feat: reorganize registry structure and fix Storybook imports
- Move layout components from registry/layout to registry/components/layout
- Move DataTable components to registry/components/data-table
- Move WorkspaceContext and WorkspaceSwitch to registry/components/layout
- Update import paths in Layout.stories.tsx and LayoutSidebarHeader.stories.tsx
- Fix DataTable.stories.tsx import paths for button component
- Rebuild registry with updated file paths
- Add CLAUDE.md with project documentation

This reorganization improves component discoverability and follows
a more consistent structure where all components are under
registry/components organized by feature.
