+++ public/r/DataTable.json
@@ -24,7 +24,7 @@
   ],
   "files": [
     {
-      "path": "registry/components/DataTable.tsx",
+      "path": "registry/components/data-table/DataTable.tsx",
       "content": "import {\n  ColumnFiltersState,\n  getCoreRowModel,\n  PaginationState,\n  SortingState,\n  useReactTable,\n  VisibilityState,\n} from '@tanstack/react-table'\nimport * as React from 'react'\n\nimport { Fn, Q, SelectQuery } from '@jakub.knejzlik/ts-query'\nimport { DataTableColumn, mapColumn } from '@/components/data-table/DataTableColumn'\nimport { StaticDataSource } from '@/lib/data-sources'\nimport { useQuery } from '@tanstack/react-query'\nimport { DataTableContent } from '@/components/data-table/DataTableContent'\nimport { DataTableFooter } from '@/components/data-table/DataTableFooter'\nimport { DataTableProvider } from '@/components/data-table/DataTableProvider'\n\ninterface DataTableProps<RecordType> {\n  columns: DataTableColumn<RecordType, any>[]\n  datasource: StaticDataSource\n  query: SelectQuery\n}\n\nexport const DataTable = <RecordType,>({\n  columns,\n  datasource,\n  query,\n  children,\n}: React.PropsWithChildren<DataTableProps<RecordType>>) => {\n  const [sorting, setSorting] = React.useState<SortingState>([])\n  const [pagination, setPagination] = React.useState<PaginationState>({\n    pageIndex: 0,\n    pageSize: 10,\n  })\n  const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([])\n  const [columnVisibility, setColumnVisibility] = React.useState<VisibilityState>({})\n  const [rowSelection, setRowSelection] = React.useState({})\n\n  let sourceQuery = query\n\n  const countQuery = Q.select().addField(Fn.count('*'), 'count').from(sourceQuery)\n\n  if (sorting.length > 0) {\n    for (const sorter of sorting) {\n      sourceQuery = sourceQuery.orderBy(sorter.id, sorter.desc ? 'DESC' : 'ASC')\n    }\n  }\n  sourceQuery = sourceQuery.offset(pagination.pageIndex * pagination.pageSize)\n  sourceQuery = sourceQuery.limit(pagination.pageSize)\n\n  const { data, isLoading } = useQuery({\n    queryKey: ['data-table', sourceQuery.toSQL()],\n    queryFn: async () => {\n      const [rows, count] = await datasource.executeQueries([sourceQuery, countQuery])\n      return [rows, { count: count[0].count }] as [RecordType[], { count: number }]\n    },\n  })\n\n  const [_data, count] = data ?? [null, null]\n\n  const table = useReactTable({\n    data: _data ?? [],\n    columns: columns.map(mapColumn),\n    manualSorting: true,\n    manualFiltering: true,\n    manualPagination: true,\n    rowCount: count ? count.count : 0,\n    // autoResetPageIndex: true,\n    onSortingChange: setSorting,\n    onColumnFiltersChange: setColumnFilters,\n    onPaginationChange: setPagination,\n    getCoreRowModel: getCoreRowModel(),\n    onColumnVisibilityChange: setColumnVisibility,\n    onRowSelectionChange: setRowSelection,\n    enableRowPinning: true,\n    state: {\n      sorting,\n      columnFilters,\n      columnVisibility,\n      rowSelection,\n      pagination,\n      rowPinning: { bottom: [''], top: [''] },\n      columnPinning: {\n        // left: ['status'],\n        right: ['actions'],\n      },\n    },\n  })\n\n  return (\n    <DataTableProvider table={table}>\n      <div className=\"w-full\">\n        {children ?? (\n          <>\n            {/* <div className=\"flex items-center py-4\">\n              <DataTableToolbar />\n            </div> */}\n            <div className=\"rounded-sm border\">\n              <DataTableContent />\n            </div>\n\n            <div className=\"flex items-center justify-end space-x-2 py-4\">\n              <DataTableFooter />\n            </div>\n          </>\n        )}\n        <div\n          hidden={!isLoading}\n          role=\"status\"\n          className=\"absolute -translate-x-1/2 -translate-y-1/2 top-2/4 left-1/2\"\n        >\n          <svg\n            aria-hidden=\"true\"\n            className=\"w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600\"\n            viewBox=\"0 0 100 101\"\n            fill=\"none\"\n            xmlns=\"http://www.w3.org/2000/svg\"\n          >\n            <path\n              d=\"M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z\"\n              fill=\"currentColor\"\n            />\n            <path\n              d=\"M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z\"\n              fill=\"currentFill\"\n            />\n          </svg>\n          <span className=\"sr-only\">Loading...</span>\n        </div>\n      </div>\n    </DataTableProvider>\n  )\n}\n",
       "type": "registry:component"
     },
