+++ registry/components/data-table/DataTable.stories.tsx
@@ -5,13 +5,13 @@ import { Q } from '@jakub.knejzlik/ts-query'
 import type { Meta, StoryObj } from '@storybook/react'
 import { PencilIcon, Trash } from 'lucide-react'
 import { DataTable } from './DataTable'
-import { Button } from './ui/button'
-import { DataTableColumn } from './data-table/DataTableColumn'
-import { DataTableToolbar } from './data-table/DataTableToolbar'
-import { DataTableContent } from './data-table/DataTableContent'
-import { DataTablePagination } from './data-table/DataTablePagination'
-import { DataTableFooter } from './data-table/DataTableFooter'
-import { DataTablePageSwitcher } from './data-table/DataTablePageSwitcher'
+import { Button } from '../ui/button'
+import { DataTableColumn } from './DataTableColumn'
+import { DataTableToolbar } from './DataTableToolbar'
+import { DataTableContent } from './DataTableContent'
+import { DataTablePagination } from './DataTablePagination'
+import { DataTableFooter } from './DataTableFooter'
+import { DataTablePageSwitcher } from './DataTablePageSwitcher'
 
 const meta: Meta<typeof DataTable> = {
   component: DataTable,
