+++ registry/components/layout/LayoutSidebarFooter.tsx
